hotfix

+++ server/src/controllers/cardController.ts
@@ -17,6 +17,14 @@ export const getCardsBySetId = async (req: Request, res: Response) => {
       return res.status(404).json({ message: 'Card set not found.' });
     }
 
+    // If the actual cards and the references are out of sync, clean up the references
+    if (cardSet.cards.length !== cardSet.cards.filter(card => card !== null).length) {
+      CardSet.updateOne(
+        { _id: setId },
+        { cards: cardSet.cards.filter(card => card !== null).map(card => card._id) }
+      ).catch((error: any) => console.error("Error cleaning up card references:", error));
+    }
+
     // If the set is found, return the populated 'cards' array
     res.status(200).json(cardSet);
     
@@ -28,17 +36,18 @@ export const getCardsBySetId = async (req: Request, res: Response) => {
 
 // // Get a single card by ID from a specific set
 export const getCardById = async (req: Request, res: Response) => {
+  try {
     const { cardId } = req.params;
-    try {
-      const card = await Card.findById(cardId);
-      if (!card) {
-        return res.status(404).json({ error: 'Card not found' });
-      }
-      res.json(card);
-    } catch (error: any) {
-      console.error(error.message);
-      res.status(500).json({ message: 'Server Error' });
+    const card = await Card.findById(cardId);
+    if (!card) {
+      return res.status(404).json({ error: 'Card not found' });
     }
+    res.json(card);
+
+  } catch (error: any) {
+    console.error(error.message);
+    res.status(500).json({ message: 'Server Error' });
+  }
 };
 
 // // Create a new card in a specific set
@@ -46,6 +55,7 @@ export const createCard = async (req: Request, res: Response) => {
   try {
     const { genre, question, correctAnswer, incorrectAnswers, difficulty } = req.body;
     const newCard = new Card({
+      user_id: req.body.userId,
       genre,
       question,
       correctAnswer,
@@ -54,6 +64,7 @@ export const createCard = async (req: Request, res: Response) => {
     });
     const savedCard = await newCard.save();
     res.status(201).json(savedCard);
+
   } catch (error: any) {
     console.error(error.message);
     res.status(500).json({ message: 'Server Error' });
@@ -63,14 +74,22 @@ export const createCard = async (req: Request, res: Response) => {
 // // Update an existing card
 export const updateCard = async (req: Request, res: Response) => {
   try {
-    // get cardId from params
+    // Get cardId from params
     const { cardId } = req.params;
-    // get card from db
+    const { userId } = req.body;
+
+    // Get card from db
     const card = await Card.findById(cardId);
     if (!card) {
       return res.status(404).json({ message: 'Card not found' });
     }
-    // update card fields
+
+    // Verify the user is the owner of the card
+    if (card.user_id.toString() !== userId) {
+      return res.status(403).json({ message: 'Forbidden: You do not have permission to update this card' });
+    }
+
+    // Update card fields
     const { genre, question, correctAnswer, incorrectAnswers, difficulty } = req.body;
     if (genre) card.genre = genre;
     if (question) card.question = question;
@@ -87,16 +106,37 @@ export const updateCard = async (req: Request, res: Response) => {
 };
 
 // // Delete a card from a set
-export const deleteCard = (req: Request, res: Response) => {
+export const deleteCard = async (req: Request, res: Response) => {
   try {
     const { cardId } = req.params;
+    const { userId } = req.body;
+
     // Find the card by its ID and delete it
+    const card = await Card.findById(cardId);
+
+    // Verify the card exists
+    if (!card) {
+      return res.status(404).json({ message: 'Card not found' });
+    }
+
+    // Verify the user is the owner of the card
+    if (card.user_id.toString() !== userId) {
+      return res.status(403).json({ message: 'Forbidden: You do not have permission to delete this card' });
+    }
+
+    // Delete the card
     Card.findByIdAndDelete(cardId)
       .then(() => res.status(204).send())
       .catch((error: any) => {
         console.error(error.message);
         res.status(500).json({ message: 'Server Error' });
       });
+
+    // Delete the card from any CardSets that contain it
+    await CardSet.updateMany(
+      { cards: cardId },
+      { $pull: { cards: cardId } }
+    );
   } catch (error: any) {
     console.error(error.message);
     res.status(500).json({ message: 'Server Error' });
